organzing changes for easier access
